+++ go/fennel/lib/clock/clock.go
@@ -3,13 +3,14 @@ package clock
 import "time"
 
 type Clock interface {
-	Now() int64
+	// TODO(REX-1157): Move back to uint64
+	Now() uint32
 }
 
 type Unix struct{}
 
 var _ Clock = Unix{}
 
-func (u Unix) Now() int64 {
-	return time.Now().Unix()
+func (u Unix) Now() uint32 {
+	return uint32(time.Now().Unix())
 }
