+++ go/fennel/model/counter/rate.go
@@ -43,7 +43,7 @@ func NewRate(tr tier.Tier, aggId ftypes.AggId, durations []uint64, normalize boo
 		aggId:     aggId,
 		Durations: durations,
 		Normalize: normalize,
-		Bucketizer: fixedWidthBucketizer{map[ftypes.Window]uint64{
+		Bucketizer: fixedWidthBucketizer{map[ftypes.Window]uint32{
 			ftypes.Window_MINUTE: 6,
 			ftypes.Window_DAY:    1,
 		}, true},
