+++ go/fennel/test/clock_test.go
@@ -12,9 +12,9 @@ func TestUnix_Now(t *testing.T) {
 
 	clock := &FakeClock{}
 	tier.Clock = clock
-	assert.Equal(t, int64(0), tier.Clock.Now())
+	assert.Equal(t, uint32(0), tier.Clock.Now())
 	clock.Set(123)
-	assert.Equal(t, int64(123), tier.Clock.Now())
+	assert.Equal(t, uint32(123), tier.Clock.Now())
 	clock.Set(12321)
-	assert.Equal(t, int64(12321), tier.Clock.Now())
+	assert.Equal(t, uint32(12321), tier.Clock.Now())
 }
